// src/frontend/src/api/generated/types.gen.ts
export type AppState_Output = {
  objectTypeColors?: {
    [key: string]: string;
  } | null;
  objectTypeClasses?: {
    [key: string]: "handling_unit" | "resource";
  } | null;
  attributeUnits?: Array<
    EventAttributeDefinition | ObjectAttributeDefinition
  > | null;
  emissionAttributes?: Array<
    EventAttributeDefinition | ObjectAttributeDefinition
  > | null;
  emissionRules?: Array<
    EventEmissionRule_Output | E2OEmissionRule_Output
  > | null;
  objectAllocationConfig?: ObjectAllocationConfig | null;
};

export type Arc = {
  source: string;
  target: string;
  double: boolean;
};

export type BaseResponse = {
  session: string;
  route: string;
  state: string;
  status?: number;
  msg?: string | null;
  appState?: AppState_Output | null;
  emissions?: ProcessEmissions | null;
};
// ...
export type E2OEmissionRule_Output = {
  index: number;
  name?: string | null;
  activity: string;
  factor: LocalEmissionFactor_Output | ClimatiqEmissionFactor_Output;
  type: "E2OEmissionRule";
  objectType: string;
  qualifier?: string | null;
  readonly defaultName: string;
};

export type EmissionsState = {
  hasImportedEmissions?: boolean;
  hasRuleBasedEmissions?: boolean;
  hasObjectEmissions: boolean;
  readonly hasEmissions: boolean;
};

export type EventAttributeDefinition = {
  target: "event";
  name: string;
  unit?: {
    /**
     * Unit symbol
     */
    symbol: string | null;
    /**
     * Long name of the unit
     */
    name: string | null;
    /**
     * The unit dimensionality, compatible with the `pint` python package
     */
    dim: {
      [key: string]: number;
    };
  } | null;
  activity: string;
};

export type EventEmissionRule_Input = {
  index: number;
  name?: string | null;
  activity: string;
  factor: LocalEmissionFactor_Input | ClimatiqEmissionFactor_Input;
  /**
   * Disables validation of the result unit being a weight.
   */
  ignoreUnit?: boolean;
  type: "EventEmissionRule";
};

export type EventEmissionRule_Output = {
  index: number;
  name?: string | null;
  activity: string;
  factor: LocalEmissionFactor_Output | ClimatiqEmissionFactor_Output;
  type: "EventEmissionRule";
  readonly defaultName: string;
};
// ...
export type NumberStats = {
  empty: boolean;
  count: number;
  sum: number;
  mean: number;
  min: number;
  median: number;
  max: number;
  nonzero: number;
};

export type OCPN = {
  objectTypes: Array<string>;
  dotString: string | null;
  structure: OcpnStructure;
};

export type ObjectAllocationConfig = {
  targetObjectTypes: Array<string>;
  rule: "AllTargets" | "ParticipatingTargets" | "ClosestTargets";
  graphMode?: "full" | "HU" | null;
  removeOtypeLoops?: boolean | null;
  maxDistance?: number | false | null;
};
// ...
export type ObjectAttributeDefinition = {
  target: "object";
  name: string;
  unit?: {
    /**
     * Unit symbol
     */
    symbol: string | null;
    /**
     * Long name of the unit
     */
    name: string | null;
    /**
     * The unit dimensionality, compatible with the `pint` python package
     */
    dim: {
      [key: string]: number;
    };
  } | null;
  objectType: string;
  dynamic: boolean;
};
// ...
export type ProcessEmissions = {
  /**
   * A unit (physical, currency, ...), represented by a symbol, name, and dimensionality.
   */
  unit: {
    /**
     * Unit symbol
     */
    symbol: string | null;
    /**
     * Long name of the unit
     */
    name: string | null;
    /**
     * The unit dimensionality, compatible with the `pint` python package
     */
    dim: {
      [key: string]: number;
    };
  };
  apiState?: string | null;
  readonly overallEmissions: number;
  readonly overallImportedEmissions: number;
  readonly overallRuleBasedEmissions: number;
  readonly activityEmissions: {
    [key: string]: NumberStats;
  };
  readonly state: EmissionsState;
};
// ...
export type TaskResponse = {
  id: string;
  route: string;
  taskState: TaskState;
  percentage?: number | null;
  msg?: string | null;
  result?: BaseResponse | null;
};

export type TaskState =
  | "PENDING"
  | "STARTED"
  | "PROGRESS"
  | "SUCCESS"
  | "FAILURE"
  | "RETRY";

export type TaskStatusResponse = {
  session: string;
  route: string;
  state: string;
  status?: number;
  msg?: string | null;
  appState?: AppState_Output | null;
  emissions?: ProcessEmissions | null;
  task: TaskResponse;
};